+++ packages/fleur-benchmarks/jest.config.js
@@ -3,6 +3,7 @@ const { join } = require('path')
 module.exports = {
   ...require('../../jest.config'),
   moduleFileExtensions: ['ts', 'js'],
+  setupFilesAfterEnv: ['<rootDir>/src/setup.ts'],
   globals: {
     'ts-jest': {
       tsConfigFile: 'tsconfig.json',