+++ packages/fleur-benchmarks/src/setup.ts
@@ -1,0 +1,1 @@
+global.console.error = jest.fn()