+++ packages/fleur-react/jest.config.js
@@ -3,5 +3,4 @@ const { join } = require('path')
 module.exports = {
   ...require('../../jest.config'),
   moduleFileExtensions: ['ts', 'tsx', 'js'],
-  setupFiles: [join(__dirname, 'test_lib/enzyme.setup.ts')],
 }