+++ packages/fleur-react/src/index.ts
@@ -6,3 +6,4 @@ export {
 export { default as connectToStores, StoreGetter } from './connectToStores'
 export { useComponentContext } from './useComponentContext'
 export { useStore } from './useStore'
+export { FleurContext } from './ComponentContextProvider'