+++ packages/fleur-react/src/ssr.spec.tsx
@@ -8,6 +8,8 @@ import * as request from 'request-promise'
 
 import { ContextProp, createElementWithContext } from '.'
 import connectToStores from './connectToStores'
+import { useStore } from './useStore'
+import { FleurContext } from './ComponentContextProvider'
 
 describe('Sever side rendering', () => {
   let server: Server
@@ -15,83 +17,71 @@ describe('Sever side rendering', () => {
 
   beforeAll(() => {
     const increaseIdent = action<{ increase: number }>()
+
     const increaseOp = operation((ctx, { increase }: { increase: number }) => {
       ctx.dispatch(increaseIdent, { increase })
     })
 
     class TestStore extends Store {
       public static storeName = 'TestStore'
-
       protected state = { count: 0 }
-
       private increase = listen(increaseIdent, ({ increase }) => {
         this.updateWith(d => (d.count += increase))
       })
-
       public getCount() {
         return this.state.count
       }
     }
 
-    const Component = connectToStores([TestStore], getStore => ({
-      count: getStore(TestStore).getCount(),
-    }))(
-      class extends React.PureComponent<{ count: number } & ContextProp> {
-        public render() {
-          return React.createElement(
-            'div',
-            {},
-            `Your count ${this.props.count}`,
-          )
-        }
-      },
-    )
+    // const Component = connectToStores([TestStore], getStore => ({
+    //   count: getStore(TestStore).getCount(),
+    // }))((props: { count: number }) => <div>{`Your count ${props.count}`}</div>)
 
-    app = new Fleur()
-    app.registerStore(TestStore)
+    const Component = () => {
+      const { count } = useStore([TestStore], getStore => ({
+        count: getStore(TestStore).getCount(),
+      }))
+
+      return <div>{`Your count ${count}`}</div>
+    }
+
+    app = new Fleur({ stores: [TestStore] })
 
     const serverApp = express()
-
     serverApp.get('/', async (req, res) => {
       try {
         const context = app.createContext()
-        context.getStore(TestStore)
+
         await context.executeOperation(increaseOp, {
           increase: parseInt(req.query.amount, 10),
         })
 
         res.write(
           ReactDOMServer.renderToString(
-            React.createElement(
-              'html',
-              {},
-              React.createElement(React.Fragment as any, {
-                children: [
-                  React.createElement(
-                    'head',
-                    {},
-                    React.createElement('script', {
-                      'data-state': JSON.stringify(context.dehydrate()),
-                    }),
-                  ),
-                  React.createElement(
-                    'body',
-                    {},
-                    createElementWithContext(context, Component, {}),
-                  ),
-                ],
-              }),
-            ),
+            <html>
+              <head>
+                <script data-state={JSON.stringify(context.dehydrate())} />
+              </head>
+              <body>
+                <FleurContext value={context}>
+                  <Component />
+                </FleurContext>
+              </body>
+            </html>,
           ),
         )
+
         res.end()
       } catch (e) {
         server && server.close()
         throw e
       }
     })
-
     server = serverApp.listen(31987)
+  })
+
+  afterAll(() => {
+    server && server.close()
   })
 
   it('test', async () => {
@@ -103,9 +93,9 @@ describe('Sever side rendering', () => {
           .load(res1)('script')
           .attr('data-state'),
       )
+
       expect(res1).toContain('<div>Your count 10</div>')
       expect(dehydratedState1).toEqual({ stores: { TestStore: { count: 10 } } })
-
       const clientContext1 = app.createContext()
       clientContext1.rehydrate(dehydratedState1)
       expect(clientContext1.getStore('TestStore').state).toEqual({ count: 10 })
@@ -117,17 +107,14 @@ describe('Sever side rendering', () => {
           .load(res2)('script')
           .attr('data-state'),
       )
+
       expect(res2).toContain('<div>Your count 20</div>')
       expect(dehydratedState2).toEqual({ stores: { TestStore: { count: 20 } } })
-
       const clientContext2 = app.createContext()
       clientContext2.rehydrate(dehydratedState2)
       expect(clientContext2.getStore('TestStore').state).toEqual({ count: 20 })
     } catch (e) {
-      server && server.close()
       throw e
     }
-
-    server && server.close()
   })
 })